Ignore error during installation of libwebkit2gtk
During the installation of libwebkit2gtk-4.0-dev the apparmor-profile-load script called by the post installation script of lxc-common fails due to the missing systemd-detect-virt binary, which is not present in trusty but expected to be there as we are using packages from xenial. Consequently the whole installation command is considered as failed. However the installation of libwebkit2gtk-4.0-dev is actually successful, so we can ignore this error.